requirements.txt
python-bitcoinrpc==1.0
requests==2.31.0
starknet.py==0.23.0